requirements.txt
pydub
requests
Flask
python-dotenv